requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2==2.8.6
djangorestframework_simplejwt==4.8.0
